+++ view/next-project/src/components/sponsoractivities/AddBlankReceiptModal.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { SponsorActivityView } from '@/type/common';
 import { PreviewPDF, createSponsorActivitiesPDF } from '@/utils/createSponsorActivitiesReceiptsPDF';
-import { getToday } from '@/utils/dateUtils';
+import { getToday } from '@/utils/dateConverter';
 import { Modal, Input, PrimaryButton, CloseButton } from '@components/common';
 
 interface ModalProps {
