+++ view/next-project/src/components/sponsoractivities/AddPdfDetailModal.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { OpenEditInvoiceModalButton } from './index';
 import { createSponsorActivitiesPDF } from '@/utils/createSponsorActivitiesInvoicesPDF';
 import { PreviewPDF } from '@/utils/createSponsorActivitiesInvoicesPDF';
-import { getToday } from '@/utils/dateUtils';
+import { getToday } from '@/utils/dateConverter';
 import { CloseButton, Input, Modal, PrimaryButton } from '@components/common';
 import {
   SponsorActivityView,
